elf: Detect if Elf is PIE
It is not sufficient to determine whether Elf is ET_DYN or not to be a
dynamic library in bcc_elf_is_shared_obj() function, because the executable
Elf of the PIE type is also ET_DYN.

For example, on Fedora 41, /usr/bin/bash is PIE:

    $ readelf -h /usr/bin/bash
    ELF Header:
      ...
      Type:                     DYN (Position-Independent Executable file)
      ...

bcc_elf_is_shared_obj() should not return 'true' for /usr/bin/bash.

This commit add function bcc_elf_is_pie() and export it in bcc_elf.h.

At the same time, fix test py_test_usdt3.

Signed-off-by: Rong Tao <rongtao@cestc.cn>
Signed-off-by: Jiang Guirong <2386499836@qq.com>
Signed-off-by: Monu Lakshkar <mlakshkar@newrelic.com>