libbpf-tools: opensnoop: add full-path argument -F (#5323)
This patch implements full-path, which tracks the full path of an open file
when the -F,--full-path parameter is specified.  The implementation differs
from patch [1] in that full path data is transmitted at one event rather than
in multi events.

For example:

    $ cd /var/tmp/
    $ touch a.out

    $ sudo ./opensnoop -F
    PID    COMM              FD ERR PATH
    106779 touch              3   0 /var/tmp/a.out
                                    ^^^^^^^^^^^^^^

Because the transmitted data is very large, it is not possible to define
the variable "struct event = {};" in stack, so the event data is reserved
in the way of bpf_buffer+heap.

Link: https://github.com/iovisor/bcc/commit/c110a4dd0c8f [1]

Signed-off-by: Rong Tao <rongtao@cestc.cn>
Signed-off-by: Monu Lakshkar <mlakshkar@newrelic.com>